+++ lib/services/message_service.dart
@@ -2,6 +2,7 @@ import 'dart:convert';
 
 import 'package:http/http.dart' as http;
 import 'package:rocket_chat_connector_flutter/exceptions/exception.dart';
+import 'package:rocket_chat_connector_flutter/models/authentication.dart';
 import 'package:rocket_chat_connector_flutter/models/new/message_new.dart';
 import 'package:rocket_chat_connector_flutter/models/response/message_new_response.dart';
 import 'package:rocket_chat_connector_flutter/services/http_service.dart';
@@ -11,9 +12,13 @@ class MessageService {
 
   MessageService(this._httpService);
 
-  Future<MessageNewResponse> postMessage(MessageNew message) async {
+  Future<MessageNewResponse> postMessage(
+      MessageNew message, Authentication authentication) async {
     http.Response response = await _httpService.post(
-        '/api/v1/chat.postMessage', jsonEncode(message.toMap()));
+      '/api/v1/chat.postMessage',
+      jsonEncode(message.toMap()),
+      authentication,
+    );
 
     if (response?.statusCode == 200) {
       if (response?.body?.isNotEmpty == true) {