+++ lib/services/user_service.dart
@@ -2,6 +2,7 @@ import 'dart:convert';
 
 import 'package:http/http.dart' as http;
 import 'package:rocket_chat_connector_flutter/exceptions/exception.dart';
+import 'package:rocket_chat_connector_flutter/models/authentication.dart';
 import 'package:rocket_chat_connector_flutter/models/new/user_new.dart';
 import 'package:rocket_chat_connector_flutter/models/user.dart';
 import 'package:rocket_chat_connector_flutter/services/http_service.dart';
@@ -11,9 +12,12 @@ class UserService {
 
   UserService(this._httpService);
 
-  Future<User> create(UserNew userNew) async {
+  Future<User> create(UserNew userNew, Authentication authentication) async {
     http.Response response = await _httpService.post(
-        '/api/v1/users.create', jsonEncode(userNew.toMap()));
+      '/api/v1/users.create',
+      jsonEncode(userNew.toMap()),
+      authentication,
+    );
 
     if (response?.statusCode == 200) {
       if (response?.body?.isNotEmpty == true) {