+++ test/services/user_service_test.dart
@@ -3,6 +3,7 @@ import 'dart:convert';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:http/http.dart';
 import 'package:mockito/mockito.dart';
+import 'package:rocket_chat_connector_flutter/models/authentication.dart';
 import 'package:rocket_chat_connector_flutter/models/new/user_new.dart';
 import 'package:rocket_chat_connector_flutter/models/user.dart';
 import 'package:rocket_chat_connector_flutter/services/http_service.dart';
@@ -16,6 +17,7 @@ class HttpServiceMock extends Mock implements HttpService {}
 void main() {
   HttpService httpServiceMock;
   UserService userService;
+  Authentication authenticationMock = new Authentication();
 
   UserNew userNew = UserNewData.getById(1);
 
@@ -28,10 +30,12 @@ void main() {
     Response response =
         Response(jsonEncode(UserData.getMapById("BsNr28znDkG8aeo7W")), 200);
     when(httpServiceMock.post(
-            "/api/v1/users.create", jsonEncode(userNew.toMap())))
-        .thenAnswer((_) => Future(() => response));
+      "/api/v1/users.create",
+      jsonEncode(userNew.toMap()),
+      authenticationMock,
+    )).thenAnswer((_) => Future(() => response));
 
-    User user = await userService.create(userNew);
+    User user = await userService.create(userNew, authenticationMock);
     expect(user.success, true);
   });
 }